+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- Serial descriptor builder based on 
 - Rows to envelope converter
 - Validation for MetaSpec
 - Detailed Meta validation result
@@ -23,6 +24,7 @@
 ### Deprecated
 
 ### Removed
+- String name pattern is replaced by a class-based name pattern 
 
 ### Fixed
 - Meta validation for multiple list values with `allowedValues` and `multiple` turned on.